Added some values to Competitor, renamed lastTime, startTime and finishTime to elapsedTime, startClock and finishClock respectively. This is to make their purpose a bit clearer.
elapsedMinutes and Hours have been made Class wide integers so I can access them in another function. This is to set the elapsedTime for a competitor in a way that is readable.

Also added some functionality to the stop button, namely stopping individual competitors to display their finish time and for how long they went.